Move relevant utility classes to maps module
