+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -66,9 +66,7 @@ import org.odk.collect.android.gdrive.GoogleAccountCredentialGoogleAccountPicker
 import org.odk.collect.android.gdrive.GoogleAccountPicker;
 import org.odk.collect.android.gdrive.GoogleAccountsManager;
 import org.odk.collect.android.gdrive.GoogleApiProvider;
-import org.odk.collect.android.geo.DirectoryReferenceLayerRepository;
 import org.odk.collect.android.geo.MapProvider;
-import org.odk.collect.android.geo.ReferenceLayerRepository;
 import org.odk.collect.android.instancemanagement.InstanceAutoSender;
 import org.odk.collect.android.itemsets.FastExternalItemsetsRepository;
 import org.odk.collect.android.logic.PropertyManager;
@@ -123,11 +121,13 @@ import org.odk.collect.async.Scheduler;
 import org.odk.collect.audiorecorder.recording.AudioRecorder;
 import org.odk.collect.audiorecorder.recording.AudioRecorderFactory;
 import org.odk.collect.forms.FormsRepository;
-import org.odk.collect.maps.MapFragmentFactory;
 import org.odk.collect.imageloader.GlideImageLoader;
 import org.odk.collect.imageloader.ImageLoader;
 import org.odk.collect.location.LocationClient;
 import org.odk.collect.location.LocationClientProvider;
+import org.odk.collect.maps.layers.DirectoryReferenceLayerRepository;
+import org.odk.collect.maps.MapFragmentFactory;
+import org.odk.collect.maps.layers.ReferenceLayerRepository;
 import org.odk.collect.permissions.PermissionsChecker;
 import org.odk.collect.permissions.PermissionsProvider;
 import org.odk.collect.projects.ProjectsRepository;
