+++ collect_app/src/main/java/org/odk/collect/android/preferences/screens/MapsPreferencesFragment.kt
@@ -23,13 +23,13 @@ import androidx.preference.PreferenceCategory
 import org.odk.collect.android.R
 import org.odk.collect.android.geo.MapConfigurator
 import org.odk.collect.android.geo.MapProvider
-import org.odk.collect.android.geo.ReferenceLayerRepository
 import org.odk.collect.android.injection.DaggerUtils
 import org.odk.collect.android.preferences.CaptionedListPreference
 import org.odk.collect.android.preferences.PrefUtils
 import org.odk.collect.android.preferences.dialogs.ReferenceLayerPreferenceDialog
 import org.odk.collect.android.preferences.screens.ReferenceLayerPreferenceUtils.populateReferenceLayerPref
 import org.odk.collect.androidshared.ui.multiclicksafe.MultiClickGuard.allowClick
+import org.odk.collect.maps.layers.ReferenceLayerRepository
 import org.odk.collect.settings.keys.ProjectKeys.CATEGORY_BASEMAP
 import org.odk.collect.settings.keys.ProjectKeys.KEY_BASEMAP_SOURCE
 import java.io.File
