+++ download-robolectric-deps.sh
@@ -12,3 +12,5 @@ mkdir -p location/src/test/resources && cp robolectric-deps.properties "$_"
 mkdir -p androidshared/src/test/resources && cp robolectric-deps.properties "$_"
 mkdir -p geo/src/test/resources && cp robolectric-deps.properties "$_"
 mkdir -p permissions/src/test/resources && cp robolectric-deps.properties "$_"
+mkdir -p settings/src/test/resources && cp robolectric-deps.properties "$_"
+mkdir -p maps/src/test/resources && cp robolectric-deps.properties "$_"
