+++ maps/src/main/java/org/odk/collect/maps/layers/ReferenceLayerRepository.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.geo
+package org.odk.collect.maps.layers
 
 import java.io.File
